+++ cuda/src/main/java9/module-info.java
@@ -15,6 +15,7 @@ module org.bytedeco.cuda {
 //  exports org.bytedeco.cuda.nppicom;
   exports org.bytedeco.cuda.nppig;
 //  exports org.bytedeco.cuda.nvgraph;
+  exports org.bytedeco.cuda.nvjpeg;
   exports org.bytedeco.cuda.nvml;
   exports org.bytedeco.cuda.nvrtc;
   exports org.bytedeco.cuda.nvToolsExt;